Creating netlify.toml file
The PR created for this branch started failing. Things are working
locally, so I will assume that the issue is a configuration matter
on the netlify side. Not sure what changed, but Netlify is unable to
find cypress now.

This patch adds a netlify.toml file, and support for cypress plugin. It
also changes the build behavior to not run the cypress tests. The idea,
is that with the cypress plugin doing them on the site after deploy they
won't be needed.